Removed SequenceManager is there was only ever one sequence
